Delete function rooms works & moved idleChecker to new class

+++ bot.php
@@ -1,5 +1,6 @@
 <?php
-unlink("./botbrain");
+if(file_exists("./botbrain"))
+  unlink("./botbrain");
 require_once "bot.class.php";
 
 $bot = new tsbot();
